spot light shadow cube map

+++ CGWork/CG_Object.h
@@ -114,7 +114,7 @@ namespace CG
 		void LookAt(vec4& eye, vec4& at, vec4& up);
 		void Ortho(double left, double right, double bottom, double top, double zNear, double zFar);
 		void Perspective(double fovY, double aspectRatio, double zNear, double zFar);
-		static mat4 ToScreenSpace(double width, double height); // default volume XY to screen space XY
+		static mat4 ToScreenSpace(int width, int height); // default volume XY to screen space XY
 		bool IsInsideFrustum(vec4& p) const;
 	};
 }
